Renamed solvers folder to puzzles
