+++ GridPuzzleSolverUnitTests/Puzzles/Kakuro/Parser/KakuroParserUnitTests.cs
@@ -2,7 +2,7 @@
 using NUnit.Framework;
 using System.Text;
 
-namespace GridPuzzleSolver.Solvers.KakuroSolver.Parser.UnitTests
+namespace GridPuzzleSolver.Puzzles.Kakuro.Parser.UnitTests
 {
     [TestFixture]
     public class KakuroParserUnitTests
