+++ GridPuzzleSolverUnitTests/Puzzles/Sudoku/SudokuPuzzleUnitTests.cs
@@ -1,7 +1,7 @@
-﻿using GridPuzzleSolver.Solvers.SudokuSolver.Parser;
+﻿using GridPuzzleSolver.Puzzles.Sudoku.Parser;
 using NUnit.Framework;
 
-namespace GridPuzzleSolver.Solvers.SudokuSolver.UnitTests
+namespace GridPuzzleSolver.Puzzles.Sudoku.UnitTests
 {
     [TestFixture]
     public class SudokuPuzzleUnitTests
